requirements.txt
quickbooks-python
pandas
python-dotenv
requests
openpyxl